requirements.txt
mysql-connector-python
dnspython
pandas
numpy
ensure
pytest

-e .